refactor(ui): Modularize UI components and fix tab rendering
This commit addresses the issue where tabs were not being displayed in the top bar of the application. The fix involved a significant refactoring of the UI code to improve modularity and correct the rendering logic.

The main changes are:
- The monolithic `layout.rs` has been broken down into smaller, more focused modules: `top_bar.rs`, `middle_pane.rs`, and `right_pane.rs`. Each module is now responsible for rendering its specific part of the UI.
- The rendering logic for the tabs in `top_bar.rs` has been corrected. The original issue was caused by rendering a widget with borders into a layout area that was too small to accommodate both the content and the borders. The layout constraints have been adjusted to provide sufficient space, ensuring the tabs are now visible.
- The function signatures for rendering have been corrected to take an immutable `&AppState`, adhering to best practices for the `ratatui` library where rendering functions should not mutate state.

With these changes, the tabs are now correctly displayed, and the pre-existing keybindings for creating, closing, and switching tabs are now functional.
