+++ crates/ui/src/lib.rs
@@ -1,3 +1,6 @@
 pub mod tui;
 pub mod layout;
 pub mod left_pane;
+pub mod top_bar;
+pub mod right_pane;
+pub mod middle_pane;
